Show recently visited tasks
Next to the current task, we show the 3 recent visited tasks.
The recent tasks are sorted by visit time. The most recent is on the left,
the oldest on the right. The oldest task will be removed when more than 3 recent
tasks will be displayed.
This will make navigating between tasks easier.