Add feature for storing flag in database
After a task is marked as "Solved", the user is asked to enter the flag.
The flag will be stored in the database and the user can open it using the
context menu. If no flag is provided, nothing will be overwritten.
Therefore, it is possible to solve and unsolve a task multiple times
without overwriting the current flag.

A global option "store-flag" is added to enable/disable this feature.
Default is true.
An API is introduced to query non-private settings, so it will also work
for non-admin users.
The "store-flag" setting can also be overwritten by the environment
variable CTFNOTE_STORE_FLAG.