Fix issue with env variables and private column
The environment variables did not use the new private column, which caused
an exception. Now, a new argument is introduced to toggle this option.
Also, the INSERT INTO query is removed in order to let the
PersistentConfiguration.setIfNotSet
decide whether to use the env variables (with default values if not set)
or the current stored database option.