Don't show "Show flag" when flag is null
The button is not useful if there is no flag yet, so now we hide the button
until a flag is stored. After that, the button will stay there forever,
because it can not be set to null again.
Therefore, you also know when a flag was submitted in the past but was
overwritten with something else if the button is there but the flag is not.