Add Hedgedoc authentication link
When a new account is created, CTFNote now tries to register the user to
Hedgedoc. After that, the user will be automatically authenticated to
Hedgedoc when logging in.
The slug and password of the user are used for registering.

Only new accounts are registered to Hedgedoc. However, on every log in
CTFNote tries to authenticate the user to Hedgedoc. This does not cause
an error.

The benefit of authenticating is that the cursor of the user shows the
username/slug and changes by users are shown by Hedgedoc.

The slug is used instead of the username because the slug never changes
for a user. If the username was used and it changes, then authentication
will always fail on login.

This feature can be disabled by the env variable MD_AUTH.