Move from synchronize to migrations
Migrations are a safer way of changing the database. The docs of TypeORM
states that synchronize should only be used in development. However,
currently it is used in production.

Now it has been changed to only migrations so no data loss could ever happen.
The existing migrations are fixed to reflect the current dev branch.

It could be possible that current databases are not compatibly with the
migrations. This is because both migrations already happened but not recorded in
the database. A fresh database is required to fix this issue or the user
should manually skip the migrations.