Change players-tooltip and add player-chips back
The active-players-on-this-task tooltip is now a separate component.
This is done because it is now used in two places:
1. In the top right corner of every task
2. In the drawer menu list of all tasks
(2) is newly introduced in this branch. This is added because it is also
a helpful place to show this information.
You can trigger it by hovering on the icon in the list.

Also, the player chips are added back again. I did this because I really
liked the idea to see (without hovering over every task) who is working
on what. This is very useful information so, for example, I know within
seconds which voice channel to join in Discord.
Otherwise, I would have to hover over each task to check who is doing what.
The chips are not shown in ultradense mode.

This commit also fixes two minor issues:
1. player.slug doesn't exist anymore (is undefined) and therefore
the color hash would be always the same for every player.
2. The category text was black in the drawer list. However, the color hash
often outputs dark colors, so now the font color is white.