Add Hedgedoc authentication
When a new account is created, CTFNote now tries to register the user to Hedgedoc. After that, the user will be automatically authenticated to Hedgedoc when logging in. The slug and password of the user are used for registering.

The benefit of authenticating is that the cursor of the user shows the username and changes by users are shown by Hedgedoc.

No new URL is required, because the MD create url is used for building the URL.

Follow up of https://github.com/TFNS/CTFNote/pull/32