Migration: add v1 user import
There are two breaking changes :

If a user has a password of size bigger than 255 chars, they will not be able to
connect anymore. We believe that this should not happen too often.
This happens because pg_crypto does not handle $2b$ hashes, so we need to
convert them to $2a$ that are limited to 255 chars.

If a user had CTF_EDIT but not CTF_ALL, the CTF_ALL permission will be granted.
This happens because we reworked the permissions system. It is not possible to
have guests with the CTF_EDIT permission anymore.

Also, a random color is picked. I hope you like yours.