Move solved tasks to end of list, add filter for own tasks
If a task is solved, it is moved to the bottom of the view
to prioritize the tasks that are not solved yet.

There is also a new checkbox introduced: 'Show my tasks'.
This will show only the tasks that you have assigned yourself to.

The 'Hide solved' checkbox now only makes the list shorter,
but it is still nice to have.