Fix development setup
The old hash pointed to a non-existing container.
I have updated it to the correct hash:
https://quay.io/repository/hedgedoc/hedgedoc/manifest/sha256:8df47646f2c58959921d2b7719a6f3023441bdb7435040aa3c32abe6145095f6
for both production and development.

Moreover, the api did not work due to an outdated version of ts-node and typescript.
This could be due to running a different node version locally on my machine.
However, running it in Docker now also works.