Make tasks exportable as Markdown
All tasks can now be bundled together to create one big document
of tasks that can be submitted as a writeup for a CTF.
This saves lots of time copying the tasks in a document one by one.

The user can choose to export all tasks or only the solved tasks.

Optionally, the user can supply a team name to add to the writeup since
that is often required if you want to submit it to an organization.

The export is done client side using the APIs of Hedgedoc.