Add HTB parser
This parses the tasks of the CTF website https://ctf.hackthebox.com/
The tasks can be copied from the network tab. A direct GET-request to the
endpoint does not work, since the endpoint requires authentication with
a bearer token in the Authorization header.

The number to category mapping seems to be static and comes from the
https://ctf.hackthebox.com/api/public/challengeCategories endpoint, which
is publicly accessible.

This parser also imports the challenge description. To allow this, the
`ParsedTask` type is modified to include an optional `description` field.
The object passed to `createTask` is also implicit,
to include the description in the object.

This commit also includes small code refactoring:
- The `ParsedTask` type is not duplicated anymore,
but imported from `parsers/index.ts`.
- Because of this, the `keep` field is also added to fix typing in `TaskImportDialog.vue`.
- The dialog now only hides when the importing is fully finished
to prevent confusion with an empty screen after importing (due to delay of importing).
The button displays the loading progress.