Add context menu to task dropdown
Now it is possible to edit a task directly from within the Hedgedoc view
of a task. This is especially useful when handing in the flag when
a task is solved, since you don't have to leave the Hedgedoc view anymore.

The context menu works for both the current task it displays,
as well as all the tasks in the dropdown menu.

In order to make this possible, the `solveTaskPopup`, `deleteTaskPopup` and
`editTaskPopup` have been moved from a `provide` to a global function
that can be imported. This allows the use of these functions within
every component instead of only the child components of the TaskList.