Add zip export
This adds the possibility to use a zip archive as the export format.
It will add every task as a Markdown file to the zip archive and create an
overview file which contains the same introduction as the Markdown export.
The original single-markdown export is still possible and the default.

To create the zip archive, the dependency [JSZip](https://www.npmjs.com/package/jszip) is used.

This feature was also already suggested in https://github.com/TFNS/CTFNote/pull/185#pullrequestreview-1095672551.