Upgrade packages of frontend to latest versions
The SCSS variables are from https://github.com/quasarframework/quasar/issues/15144#issue-1502838365
This fixes the ugly dark shadows in dark mode but keeps the shadows in light mode.
This was added in Quasar 2.11 but reverted using the SCSS patch.

Due to an update of ESLint, it complains about single-word components.
Since this is used in CTFNote, this rule is disabeld.

Also some small (automatic) fixes.