Add pre-build images through GHCR
Docker images of the front, api and db will be automatically build through
Github workflows and pushed to the Github Container Registry (GHCR)
when a commit is made to the main branch (or when manually triggered).

This makes it possible to install CTFNote on a VPS of 2 GB of RAM since
you don't need to build the container anymore (which requires quite
a lot of RAM). Building the container yourself is still possible
and the default setup for docker compose.

The README is updated with this change.
Futhermore, the REAMDE now points to `docker compose` instead of the
deprecated `docker-compose`.