Add last active admin functionality
The implementation is mostly based on https://github.com/DU4L/CTFNote/pull/16
but we update the timer in the database instead of the frontend.
The update is done when the user fetches the profile,
so for every first load done in a new tab in the browser.

Please note that this commit can NOT be merged before https://github.com/TFNS/CTFNote/pull/213
is merged since the migration files are now not consecutive.
The other PR contains a migration with ID 41, so this PR has a migration
with ID 42 (and this answers everything of course).
The migration will fail in this branch therefore. For testing, please
rename the file and use a throw away database.