Add initial tags implementation
Categories are now replaced by tags which are just categories
but there can be many of them. I have implemented a many-to-many
relation from tasks to tags. Deleting tasks does not exist and will probably
not be implemented.

There is a lot of commented code in this commit.
This is on purpose to store the attempts for future references.