Add public profile by column security
Now the row `lastactive` is only accessible by the admin.
The public profile is accessed everywhere and in the admin view
the whole profile is queried.

For the me() function, we have to output a valid ctfnote.profile
but we can't access the `lastactive` column, so we replace it with
NULL by default. This fixes the problem we faced in the row wise
security model.