Add dynamic channel creation for "new, started, solved"
New: channels that are just created or without a "working on it" status
and no flag registered in the database.
Started: tasks with "working on it" status.
Solved: task with flag registered in the database.

Each state has its own category.

We also dynamically create new categories and place them correctly
in order when a category is full. The Discord limit is 50, so we shouldn't
create so much more categories during a CTF.