Handle moving of tasks to other categories
As soon as you start working on a new challenge, it will be moved to
the 'started' category. If you solve a challenge, it will be moved to
the 'solved' category. Last, if you unsolve a challenge, it will be moved
back to the 'started' category.

If a category is hitting the 50 channels limit of Discord, a new category
is automatically created.