Make task import serial to fix race condition
The channel limit can be ignored due to all the tasks being processed
in parallel, thus all thinking there is still room for the channel while
it is actually already claimed by another challenge being created.

To fix this, we have to make it all serial which makes the import way
slower but correct.