API: honour DB_HOST and DB_PORT
API needs to wait for the database to be up before starting.

We do not know how long it takes: just because the container is started doesn't
mean that Postgres is ready. The workaround is to use a small script (start.sh)
that waits until a connection to the database can be done before passing the
execution to yarn.

The Dockerfile contains hardcoded values for the database's hostname and ports,
but there is no guarantee that these values are always the same. This is
especially true when trying to deploy CTFNote with a previously-existing
database.

This commit removes the hardcoded values and replaces them with DB_HOST/DB_PORT.