Drop ECSC2021/2022 parser
The parser was introduced in https://github.com/TFNS/CTFNote/pull/184
but I don't think anyone will be using this parser since the project is dead for years
and ENISA is not using this gameboard in the last couple of years.
Since it conflicts with the parsing of the other CTF formats, it is better to drop it and reintroduce it when necessary.