Parse challenges by `parser` instead of detection through `isValid`
Now we just try to parse the input with each parser, ignoring errors. Then we select the parser with the most parsed challenges,
except for some conditions with the raw parser.
The raw parser will see every newline as a new task and when pasting a beautified JSON object,
it should not be the best parser when other parsers also detect challenges.