Dynamically import Commands and Interactions based on handle style
In the future we want to be able to change the handle style of the Discord bot (see https://github.com/TFNS/CTFNote/issues/239).
Now we have an easy way of adding new handle styles which can be toggled by .env configuration.

Only the discordHooks.ts needs to be split and reorganized for this to fully work.