Fix 50+ task creation with Discord bot enabled
Since we do not support batch importing tasks, the frontend spams the backend with a lot of task creations.
Due to throttling from Discord, it fails to create the extra Discord category for 50+ challenges.
This will break the channel creation.

So the fix for this without supporting batch import, is to import with batches from 10 where each first task is done in serial and the other 9 in parallel.